requirements.txt
tensorflow-gpu==2.2.0
numpy>=1.16.2
pytest>=5.4.3